fix macos linter
